refactor: [#1231] exctract mod
